+++ test/shackle_tests.erl
@@ -54,6 +54,20 @@ shackle_call_crash_test_() ->
         fun (_) -> cleanup(?CLIENT_TCP) end,
     [fun call_crash_subtest/0]}.
 
+shackle_random_socket_test_() ->
+    otp_28({setup,
+        fun () ->
+            setup(?CLIENT_SOCKET, [
+                {pool_size, 1},
+                {pool_strategy, random}
+        ]) end,
+        fun (_) -> cleanup(?CLIENT_SOCKET) end,
+    {inparallel, [
+        fun () -> add_subtest(?CLIENT_SOCKET) end,
+        fun () -> multiply_subtest(?CLIENT_SOCKET) end,
+        fun () -> noop_subtest(?CLIENT_SOCKET) end
+    ]}}).
+
 shackle_random_ssl_test_() ->
     {setup,
         fun () ->
@@ -67,6 +81,20 @@ shackle_random_ssl_test_() ->
         fun () -> multiply_subtest(?CLIENT_SSL) end,
         fun () -> noop_subtest(?CLIENT_SSL) end
     ]}}.
+
+shackle_random_ssl_socket_test_() ->
+    otp_28({setup,
+        fun () ->
+            setup(?CLIENT_SSL_SOCKET, [
+                {pool_size, 1},
+                {pool_strategy, random}
+        ]) end,
+        fun (_) -> cleanup(?CLIENT_SSL_SOCKET) end,
+    {inparallel, [
+        fun () -> add_subtest(?CLIENT_SSL_SOCKET) end,
+        fun () -> multiply_subtest(?CLIENT_SSL_SOCKET) end,
+        fun () -> noop_subtest(?CLIENT_SSL_SOCKET) end
+    ]}}).
 
 shackle_random_tcp_test_() ->
     {setup,
@@ -98,6 +126,15 @@ shackle_random_udp_test_() ->
         fun () -> noop_subtest(?CLIENT_UDP) end
     ]}}.
 
+shackle_reconnect_socket_test_() ->
+    otp_28({setup,
+        fun () ->
+            setup(),
+            ?CLIENT_SOCKET:start()
+        end,
+        fun (_) -> cleanup(?CLIENT_SOCKET) end,
+    [fun () -> reconnect_subtest(?CLIENT_SOCKET) end]}).
+
 shackle_reconnect_ssl_test_() ->
     {setup,
         fun () ->
@@ -106,6 +143,15 @@ shackle_reconnect_ssl_test_() ->
         end,
         fun (_) -> cleanup(?CLIENT_SSL) end,
     [fun () -> reconnect_subtest(?CLIENT_SSL) end]}.
+
+shackle_reconnect_ssl_socket_test_() ->
+    otp_28({setup,
+        fun () ->
+            setup(),
+            ?CLIENT_SSL_SOCKET:start()
+        end,
+        fun (_) -> cleanup(?CLIENT_SSL_SOCKET) end,
+    [fun () -> reconnect_subtest(?CLIENT_SSL_SOCKET) end]}).
 
 shackle_reconnect_tcp_test_() ->
     {setup,
@@ -152,6 +198,17 @@ shackle_round_robin_udp_test_() ->
         fun () -> noop_subtest(?CLIENT_UDP) end
     ]}}.
 
+shackle_timeout_socket_test_() ->
+    otp_28({setup,
+        fun () ->
+            setup(?CLIENT_SOCKET, [
+                {pool_size, 1},
+                {pool_strategy, random}
+            ])
+        end,
+        fun (_) -> cleanup(?CLIENT_SOCKET) end,
+    [fun () -> timeout_subtest(?CLIENT_SOCKET) end]}).
+
 shackle_timeout_ssl_test_() ->
     {setup,
         fun () ->
@@ -162,6 +219,17 @@ shackle_timeout_ssl_test_() ->
         end,
         fun (_) -> cleanup(?CLIENT_SSL) end,
     [fun () -> timeout_subtest(?CLIENT_SSL) end]}.
+
+shackle_timeout_ssl_socket_test_() ->
+    otp_28({setup,
+        fun () ->
+            setup(?CLIENT_SSL_SOCKET, [
+                {pool_size, 1},
+                {pool_strategy, random}
+            ])
+        end,
+        fun (_) -> cleanup(?CLIENT_SSL_SOCKET) end,
+    [fun () -> timeout_subtest(?CLIENT_SSL_SOCKET) end]}).
 
 shackle_timeout_tcp_test_() ->
     {setup,
@@ -261,6 +329,14 @@ cleanup(Client) ->
     Server:stop(),
     cleanup().
 
+otp_28(Tests) ->
+    case list_to_integer(erlang:system_info(otp_release)) >= 28 of
+        true ->
+            Tests;
+        false ->
+            []
+    end.
+
 rand() ->
     shackle_utils:random(254).
 
@@ -272,7 +348,11 @@ receive_loop(N) ->
             [X | receive_loop(N - 1)]
     end.
 
+server(?CLIENT_SOCKET) ->
+    arithmetic_tcp_server;
 server(?CLIENT_SSL) ->
+    arithmetic_ssl_server;
+server(?CLIENT_SSL_SOCKET) ->
     arithmetic_ssl_server;
 server(?CLIENT_TCP) ->
     arithmetic_tcp_server;